# Flappy Bird Part 2
## Overview of the Game
> Flappy Bird is the most popular game on the Internet these days. It is a simple one-button game in which you have to navigate our beloved bird past the obstacles.
> You do this by clicking your mouse repeatedly. Each time you click, the bird flies up. Don't make it fly too high or it will die. This means you have to tap at the right moments to make the bird go as far as you can!

## Overview of the Tutorial
This tutorial consists two parts. In the second part, we will implement the game interaction.

By the end of the tutorial, you will learn:
- handle user input in the HTML5 game
- make characters respond to user input in the game
- handle game logics with collision detection
- add a scoring system to the game
- increase game difficulty level

## Step 1. Starter Code
Our starter code continues where we left last week. In addition, we include a `collision.js` file that contains the code for detecting div collision. We will not go into the details of this collision detection function.