Step B5. Adding Collisions Sadly, we have to make our bird hit things. Let’s start by adding the final piece of this, and then work backwards. In the `update()` function, let’s add an `if()` statement that checks if our bird has collided with any of the poles or hit the top or bottom of the screen. ``` // Check for collisions if (collision(bird, poles[0]) || collision(bird, poles[1]) || birdTop <= 0 || birdTop > containerHeight - bird.clientHeight) {     gameOver(); } ``` This is saying: Game Over if - The bird hits the top pole - The bird hits the bottom pole - The bird hits the top of the frame - The bird hits the bottom of the frame
We use the `collision()` function to determine whether the bird hits the poles by checking whether the bird’s boundary and the pole’s boundary are overlapping. Remember, this is a `collision()` function, so if it returns true, we’re saying, “Yes, something is colliding”.

In this session, we don’t need to implement the `collision()` function. It is provided in `collision.js`.

Let’s test this out. Open up the console for this one. When you hit the bottom of the screen, fly past the top of the screen, or hit the poles, you should see the bird stop moving! Good job.